+++ src/object.js
@@ -91,6 +91,16 @@ export let simpleDiff = (original, deltas) => {
 }
 export let simpleDiffArray = _.flow(simpleDiff, unkeyBy('field'))
 
+export let diff = (original, deltas) => {
+  let o = flattenObject(original)
+  let d = flattenObject(deltas)
+  return _.flow(
+    mapValuesIndexed((_, field) => ({ from: o[field], to: d[field] })),
+    _.omitBy(x => x.from === x.to)
+  )(_.merge(o, d))
+}
+export let diffArray = _.flow(diff, unkeyBy('field'))
+
 // A `_.pick` that mutates the object
 export let pickOn = (paths = [], obj = {}) => _.flow(
   _.keys,