+++ test/objects.spec.js
@@ -279,4 +279,121 @@ describe('Object Functions', () => {
       to: 20
     }])
   })
+  it('diff', () => {
+    expect(f.diff({
+      x: 1,
+      a: 3,
+      d: {
+        f: 6
+      },
+      price: 20,
+      notChanged: 45
+    }, {
+      x: 1,
+      a: 1,
+      b: 2,
+      c: 3,
+      d: {
+        e: 5
+      },
+      price: undefined,
+      notChanged: 45,
+      amount: 20
+    })).to.deep.equal({
+      a: {
+        from: 3,
+        to: 1
+      },
+      b: {
+        from: undefined,
+        to: 2
+      },
+      c: {
+        from: undefined,
+        to: 3
+      },
+      'd.e': {
+        from: undefined,
+        to: 5
+      },
+      'd.f': {
+        from: 6,
+        to: undefined
+      },
+      amount: {
+        from: undefined,
+        to: 20
+      },
+      price: {
+        from: 20,
+        to: undefined
+      }
+    })
+  })
+  it('diffArray', () => {
+    expect(f.diffArray({
+      x: 1,
+      a: 3,
+      d: {
+        f: 6
+      },
+      price: 20,
+      notChanged: 45,
+      collection1: [{a: 1, b: 2}],
+      collection2: [{a: 1, b: 2}]
+    }, {
+      a: 1,
+      b: 2,
+      c: 3,
+      x: 1,
+      d: {
+        e: 5
+      },
+      price: undefined,
+      amount: 20,
+      notChanged: 45,
+      collection1: [{a: 1}],
+      collection2: []
+    })).to.deep.equal([{
+      field: 'a',
+      from: 3,
+      to: 1
+    }, {
+      field: 'd.f',
+      from: 6,
+      to: undefined
+    }, {
+      field: 'price',
+      from: 20,
+      to: undefined
+    }, {
+      field: 'collection1.b',
+      from: 2,
+      to: undefined
+    }, {
+      field: 'collection2.a',
+      from: 1,
+      to: undefined
+    }, {
+      field: 'collection2.b',
+      from: 2,
+      to: undefined
+    }, {
+      field: 'b',
+      from: undefined,
+      to: 2
+    }, {
+      field: 'c',
+      from: undefined,
+      to: 3
+    }, {
+      field: 'd.e',
+      from: undefined,
+      to: 5
+    }, {
+      field: 'amount',
+      from: undefined,
+      to: 20
+    }])
+  })
 })