+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+﻿# 1.32.0
+- Added `diff` and `diffArray`, just like `simpleDiff` and
+  `simpleDiffArray`, but they also take in count removed properties.
+
 # 1.31.0
 - Add new utility for objects called `pickOn`, which works as `pick` in lodash but it mutates the object